Filled database with data from dumps and created API to interact with database. Now only shortest-path algoithm needs to be finalized
I filled the database with all the values from the dump and created an API to interact with the database from the program. I also added indexing to the database for faster lookup. This API is used in the shortest-path algorithm. The algorithm still needs to be worked on as it is still buggy and slow.